chat_bot/requirements.txt
aiogram
aiohttp
redis
python-dotenv==1.0.0